requirements.txt
matplotlib==3.8.2
networkx==3.2.1
numpy==2.2.4
PyYAML==6.0.2
torch==2.5.1+cu124
torch_geometric==2.6.1
torch_scatter==2.1.2+pt25cu124
tqdm==4.66.1
